requirements.txt
streamlit==1.39.0
openai==1.52.2
PyPDF2==3.0.1
langchain==0.3.26
langchain-core==0.3.72
chromadb==0.5.5
python-dotenv==1.0.0
pillow==10.4.0
pypika==0.48.9
yfinance==0.2.47
langchain-community==0.3.3
newsapi-python==0.2.6
protobuf==3.20.3
faiss-cpu==1.11.0
tiktoken==0.9.0
pandas==2.3.1
requests==2.32.4
matplotlib==3.10.3
numpy==1.26.4
tweepy==4.16.0
pandas-ta==0.3.14b0